+++ comments/comments.resolvers.js
@@ -1,0 +1,6 @@
+export default {
+  Comment: {
+    isMine: ({ userId }, _, { loggedInUser }) =>
+      !loggedInUser ? false : userId === loggedInUser.id,
+  },
+};