+++ comments/deleteComment/deleteComment.typeDefs.js
@@ -1,8 +1,7 @@
 import { gql } from "apollo-server-express";
 
 export default gql`
-  scalar Upload
   type Mutation {
-    uploadPhoto(file: Upload!, caption: String): Photo
+    deleteComment(id: Int!): MutationResponse!
   }
 `;