+++ photos/photos.resolvers.js
@@ -14,6 +14,9 @@ export default {
         },
       }),
     likes: ({ id }) => client.like.count({ where: { photoId: id } }),
+    comments: ({ id }) => client.comment.count({ where: { photoId: id } }),
+    isMine: ({ userId }, _, { loggedInUser }) =>
+      !loggedInUser ? false : userId === loggedInUser.id,
   },
   Hashtag: {
     photos: ({ id }, { page }) =>