+++ photos/uploadPhoto/uploadPhoto.resolvers.js
@@ -1,5 +1,6 @@
 import { GraphQLUpload } from "graphql-upload";
 import client from "../../client";
+import { uploadToS3 } from "../../shared/shared.utils";
 import { protectedResolver } from "../../users/users.utils";
 import { processHashtags } from "../photos.utils";
 
@@ -13,9 +14,11 @@ export default {
           hashtagsObj = processHashtags(caption);
         }
 
+        const fileUrl = await uploadToS3(file, loggedInUser.id, "uploads");
+
         return client.photo.create({
           data: {
-            file,
+            file: fileUrl,
             caption,
             user: {
               connect: {