+++ users/editProfile/editProfile.resolvers.js
@@ -3,6 +3,7 @@ import client from "../../client";
 import bcrypt from "bcrypt";
 import { protectedResolver } from "../users.utils";
 import { GraphQLUpload } from "graphql-upload";
+import { uploadToS3 } from "../../shared/shared.utils";
 
 export default {
   Upload: GraphQLUpload,
@@ -27,8 +28,11 @@ export default {
         }
         let avatarUrl = null;
         if (avatar) {
+          avatarUrl = await uploadToS3(avatar, loggedInUser.id, "avatars");
+          /*
           const { filename, createReadStream } = await avatar;
           const newFilename = `${loggedInUser.id}-${Date.now()}-${filename}`;
+
           const readStream = createReadStream();
           const writeStream = fs.createWriteStream(
             process.cwd() + "/uploads/" + newFilename
@@ -36,6 +40,7 @@ export default {
 
           readStream.pipe(writeStream);
           avatarUrl = `http://localhost:4000/uploads/${newFilename}`;
+          */
         }
 
         const updatedUser = await client.user.update({