+++ users/editProfile/editProfile.typeDefs.js
@@ -2,10 +2,6 @@ import { gql } from "apollo-server";
 
 export default gql`
   scalar Upload
-  type EditProfileResult {
-    ok: Boolean!
-    error: String
-  }
 
   type Mutation {
     editProfile(
@@ -16,6 +12,6 @@ export default gql`
       password: String
       bio: String
       avatar: Upload
-    ): EditProfileResult!
+    ): MutationResponse!
   }
 `;