+++ users/unfollowUser/unfollow.typeDefs.js
@@ -1,11 +1,7 @@
 import { gql } from "apollo-server-express";
 
 export default gql`
-  type unfollowUserResult {
-    ok: Boolean
-    error: String
-  }
   type Mutation {
-    unfollowUser(username: String!): unfollowUserResult!
+    unfollowUser(username: String!): MutationResponse!
   }
 `;